Extract package-index Package into its own module
